Check that again.

+++ corr-cloud/cloud/__init__.py
@@ -129,44 +129,44 @@ def crossdomain(origin=None, methods=None, headers=None,
         return update_wrapper(wrapped_function, f)
     return decorator
 
-def load_bundle(record):
-    # Include record files later.
-    memory_file = BytesIO()
-    with zipfile.ZipFile(memory_file, 'w') as zf:
+# def load_bundle(record):
+#     # Include record files later.
+#     memory_file = BytesIO()
+#     with zipfile.ZipFile(memory_file, 'w') as zf:
 
-        try:
-            bundle_buffer = StringIO()
-            # with open(record.container.image['location'], 'rb') as fh:
-            #     image_buffer.write(fh.read())
-            # res = key.get_contents_to_filename(record.container.image['location'])   
-            # s3_client = boto3.client('s3')
-            # res = s3_client.get_object(Bucket='ddsm-bucket', Key=record.container.image['location']) 
-            obj = s3.Object(bucket_name='reproforge-bundles', key=record.environment.bundle['location'])  
-            res = obj.get()      
-            bundle_buffer.write(res['Body'].read())
-            bundle_buffer.seek(0)
+#         try:
+#             bundle_buffer = StringIO()
+#             # with open(record.container.image['location'], 'rb') as fh:
+#             #     image_buffer.write(fh.read())
+#             # res = key.get_contents_to_filename(record.container.image['location'])   
+#             # s3_client = boto3.client('s3')
+#             # res = s3_client.get_object(Bucket='ddsm-bucket', Key=record.container.image['location']) 
+#             obj = s3.Object(bucket_name='reproforge-bundles', key=record.environment.bundle['location'])  
+#             res = obj.get()      
+#             bundle_buffer.write(res['Body'].read())
+#             bundle_buffer.seek(0)
 
-            data = zipfile.ZipInfo("%s"%(record.project.name, record.environment.bundle['location'].split('_')))
-            data.date_time = time.localtime(time.time())[:6]
-            data.compress_type = zipfile.ZIP_DEFLATED
-            data.external_attr |= 0777 << 16L # -rwx-rwx-rwx
-            zf.writestr(data, bundle_buffer.read())
-        except:
-            print traceback.print_exc()
+#             data = zipfile.ZipInfo("%s"%(record.project.name, record.environment.bundle['location'].split('_')))
+#             data.date_time = time.localtime(time.time())[:6]
+#             data.compress_type = zipfile.ZIP_DEFLATED
+#             data.external_attr |= 0777 << 16L # -rwx-rwx-rwx
+#             zf.writestr(data, bundle_buffer.read())
+#         except:
+#             print traceback.print_exc()
 
-        try:
-            json_buffer = StringIO()
-            json_buffer.write(record.to_json())
-            json_buffer.seek(0)
+#         try:
+#             json_buffer = StringIO()
+#             json_buffer.write(record.to_json())
+#             json_buffer.seek(0)
 
-            data = zipfile.ZipInfo("%s_%s.json"%(record.project.name, str(record.id)))
-            data.date_time = time.localtime(time.time())[:6]
-            data.compress_type = zipfile.ZIP_DEFLATED
-            data.external_attr |= 0777 << 16L # -rwx-rwx-rwx
-            zf.writestr(data, json_buffer.read())
-        except:
-            print traceback.print_exc()
-    memory_file.seek(0)
+#             data = zipfile.ZipInfo("%s_%s.json"%(record.project.name, str(record.id)))
+#             data.date_time = time.localtime(time.time())[:6]
+#             data.compress_type = zipfile.ZIP_DEFLATED
+#             data.external_attr |= 0777 << 16L # -rwx-rwx-rwx
+#             zf.writestr(data, json_buffer.read())
+#         except:
+#             print traceback.print_exc()
+#     memory_file.seek(0)
 
 
     # imz.append('record.tar', image_buffer.read()).append("record.json", json_buffer.read())
@@ -176,38 +176,45 @@ def load_bundle(record):
     return [memory_file, record.environment.bundle['location'].split("/")[-1].split(".")[0]+".zip"]
 
 def delete_project_files(project):
-    s3_files = s3.Bucket('reproforge-pictures')
-    s3_bundles = s3.Bucket('reproforge-bundles')
+    # s3_files = s3.Bucket('reproforge-pictures')
+    # s3_bundles = s3.Bucket('reproforge-bundles')
 
     from corrdb.common.models import ProjectModel
     from corrdb.common.models import RecordModel
     from corrdb.common.models import EnvironmentModel
     from corrdb.common.models import FileModel
 
+    # print s3_files
     # project resources
     for _file in project.resources:
         file_ = FileModel.objects.with_id(_file)
         if file_:
-            s3_files.delete_key(file_.storage)
-            file_.delete()
-            logStat(deleted=True, file=_file)
+            # print file_.to_json()
+            result = s3_delete_file(file_.group, file_.storage)
+            if result:
+                logStat(deleted=True, file=file_)
+                file_.delete()
 
     # project records resources
     for record in project.records:
-        for _file in record.resources:
-            file_ = FileModel.objects.with_id(_file)
-            if file_:
-                s3_files.delete_key(file_.storage)
-                file_.delete()
-                logStat(deleted=True, file=_file)
-        logStat(deleted=True, record=record)
+        result = delete_record_files(record)
+        if result:
+            logStat(deleted=True, record=record)
+            record.delete()
 
     for environment_id in project.history:
         _environment = EnvironmentModel.objects.with_id(environment_id)
         if _environment.bundle["scope"] == "local":
             s3_bundles.delete_key(_environment.bundle.location)
-        _environment.delete()
-        logStat(deleted=True, environment=_environment)
+            result = s3_delete_file('bundle', _environment.bundle.location)
+            if result:
+                logStat(deleted=True, bundle=_environment.bundle)
+                logStat(deleted=True, environment=_environment)
+                _environment.bundle.delete()
+                _environment.delete()
+        else:
+            logStat(deleted=True, environment=_environment)
+            _environment.delete()
 
 def cloud_response(code, title, content):
     import flask as fk
@@ -216,21 +223,24 @@ def cloud_response(code, title, content):
     return fk.Response(json.dumps(response, sort_keys=True, indent=4, separators=(',', ': ')), mimetype='application/json')
 
 def delete_record_files(record):
-    s3_files = s3.Bucket('reproforge-files')
+    # s3_files = s3.Bucket('reproforge-files')
 
     from corrdb.common.models import RecordModel
     from corrdb.common.models import FileModel
-
+    final_result = True
     for _file_id in record.resources:
         _file = FileModel.objects.with_id(_file_id)
-        s3_files.delete_key(_file.location)
-        _file.remove()
-        logStat(deleted=True, file=_file)
+        result = delete_record_file(_file)
+        if not result:
+            final_result = result
+    return final_result
 
 def delete_record_file(record_file):
-    s3_files = s3.Bucket('reproforge-pictures')
-
-    s3_files.delete_key(record_file.location)
+    result = s3_delete_file(record_file.group, record_file.storage)
+    if result:
+        logStat(deleted=True, file=record_file)
+        record_file.delete()
+    return result
 
 def s3_get_file(group='', key=''):
     file_buffer = StringIO()
@@ -288,67 +298,67 @@ def s3_delete_file(group='', key=''):
 
 
 
-def load_file(file_model):
+# def load_file(file_model):
 
-    file_buffer = StringIO()
-    obj = s3.Object(bucket_name='reproforge-files', key=file_model.location)  
-    res = obj.get()      
-    file_buffer.write(res['Body'].read())
-    file_buffer.seek(0)
+#     file_buffer = StringIO()
+#     obj = s3.Object(bucket_name='reproforge-files', key=file_model.location)  
+#     res = obj.get()      
+#     file_buffer.write(res['Body'].read())
+#     file_buffer.seek(0)
 
-    return [file_buffer, file_model.location.split('_')[1]]
+#     return [file_buffer, file_model.location.split('_')[1]]
 
-def load_picture(profile):
+# def load_picture(profile):
 
-    picture_buffer = StringIO()
-    obj = s3.Object(bucket_name='reproforge-pictures', key=profile.picture['location'])
-    res = obj.get()      
-    picture_buffer.write(res['Body'].read())
-    picture_buffer.seek(0)
+#     picture_buffer = StringIO()
+#     obj = s3.Object(bucket_name='reproforge-pictures', key=profile.picture['location'])
+#     res = obj.get()      
+#     picture_buffer.write(res['Body'].read())
+#     picture_buffer.seek(0)
 
-    return [picture_buffer, digest]
+#     return [picture_buffer, digest]
 
-def upload_bundle(current_user, environment, file_obj):
-    dest_filename = str(current_user.id)+"-"+str(environment.id)+"_%s"%file_obj.filename #kind is record| signature
+# def upload_bundle(current_user, environment, file_obj):
+#     dest_filename = str(current_user.id)+"-"+str(environment.id)+"_%s"%file_obj.filename #kind is record| signature
 
-    print "About to write..."
-    try:
-        s3.Bucket('reproforge-bundles').put_object(Key=str(current_user.id)+"-"+str(environment.id)+"_%s"%file_obj.filename, Body=file_obj.read())
-        environment.bundle['location'] = dest_filename
-        environment.bundle['size'] = file_obj.tell()
-        print "%s saving..."%file_obj.filename
-        environment.save()
-        return True
-    except:
-        return False
-        print traceback.print_exc()
+#     print "About to write..."
+#     try:
+#         s3.Bucket('reproforge-bundles').put_object(Key=str(current_user.id)+"-"+str(environment.id)+"_%s"%file_obj.filename, Body=file_obj.read())
+#         environment.bundle['location'] = dest_filename
+#         environment.bundle['size'] = file_obj.tell()
+#         print "%s saving..."%file_obj.filename
+#         environment.save()
+#         return True
+#     except:
+#         return False
+#         print traceback.print_exc()
 
-def upload_file(current_user, file_model, file_obj):
-    dest_filename = str(current_user.id)+"-"+str(file_model.record.id)+"_%s"%file_obj.filename #kind is record| signature
+# def upload_file(current_user, file_model, file_obj):
+#     dest_filename = str(current_user.id)+"-"+str(file_model.record.id)+"_%s"%file_obj.filename #kind is record| signature
 
-    print "About to write..."
-    try:
-        s3.Bucket('reproforge-files').put_object(Key=str(current_user.id)+"-"+str(record.id)+"_%s"%file_obj.filename, Body=file_obj.read())
-        file_model.location = dest_filename
-        file_model.size = file_obj.tell()
-        print "%s saving..."%file_obj.filename
-        file_model.save()
-        return True
-    except:
-        return False
-        print traceback.print_exc()
+#     print "About to write..."
+#     try:
+#         s3.Bucket('reproforge-files').put_object(Key=str(current_user.id)+"-"+str(record.id)+"_%s"%file_obj.filename, Body=file_obj.read())
+#         file_model.location = dest_filename
+#         file_model.size = file_obj.tell()
+#         print "%s saving..."%file_obj.filename
+#         file_model.save()
+#         return True
+#     except:
+#         return False
+#         print traceback.print_exc()
 
-def upload_picture(current_user, file_obj):
-    dest_filename = str(current_user.id) #kind is record| signature
+# def upload_picture(current_user, file_obj):
+#     dest_filename = str(current_user.id) #kind is record| signature
 
-    print "About to write..."
-    try:
-        s3.Bucket('reproforge-pictures').put_object(Key=str(current_user.id)+"."+file_obj.filename.split('.')[-1], Body=file_obj.read())
-        print "%s saving..."%file_obj.filename
-        return True
-    except:
-        return False
-        print traceback.print_exc()
+#     print "About to write..."
+#     try:
+#         s3.Bucket('reproforge-pictures').put_object(Key=str(current_user.id)+"."+file_obj.filename.split('.')[-1], Body=file_obj.read())
+#         print "%s saving..."%file_obj.filename
+#         return True
+#     except:
+#         return False
+#         print traceback.print_exc()
 
 def logTraffic(endpoint=''):
     # created_at=datetime.datetime.utcnow()
@@ -648,6 +658,9 @@ CLOUD_URL = '/cloud/v{0}'.format(CLOUD_VERSION)
 
 VIEW_HOST = app.config['VIEW_SETTINGS']['host']
 VIEW_PORT = app.config['VIEW_SETTINGS']['port']
+
+API_HOST = app.config['VIEW_SETTINGS']['host']
+API_PORT = app.config['VIEW_SETTINGS']['port']
 
 from . import views
 from corrdb.common import models
